Refactor setup.py to improve package installation process
