+++ setup.py
@@ -11,7 +11,7 @@ setup(
     long_description=open("README.md", encoding="utf-8").read(),
     long_description_content_type="text/markdown",
     url="https://github.com/ec-intl/tcolors",
-    packages=find_packages(where="src"),
+    packages=["tcolors"],
     package_dir={"": "src"},
     license="Apache License 2.0",
     classifiers=[
@@ -20,7 +20,4 @@ setup(
         "Operating System :: OS Independent",
     ],
     python_requires=">=3.6",
-    install_requires=open("requirements.txt", encoding="utf-8")
-    .read()
-    .splitlines(),
 )
